requirements.txt
# Core Data Processing (Python 3.13 compatible)
pandas>=2.2.0
numpy>=1.26.0
polars>=0.20.3

# Machine Learning
scikit-learn>=1.4.0
xgboost>=2.0.3
# lightgbm>=4.2.0  # Optional - may need compilation
# tensorflow>=2.15.0  # Optional - requires specific setup

# Time Series Analysis
prophet>=1.1.5
statsmodels>=0.14.1
# pmdarima>=2.0.4  # Optional - may need compilation

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Dashboard (Essential)
streamlit>=1.31.0
# streamlit-aggrid>=0.3.4  # Optional
# streamlit-option-menu>=0.3.6  # Optional

# Data Quality & Anomaly Detection
pyod>=1.1.0
# isolation-forest>=0.2.1  # Optional - use sklearn instead

# Geospatial Analysis (Optional)
# geopandas>=0.14.2
# folium>=0.15.1
# shapely>=2.0.2

# Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Explainable AI (Optional)
# shap>=0.44.0
# lime>=0.2.0.1

# Database
sqlalchemy>=2.0.25

# Reporting
openpyxl>=3.1.0
xlsxwriter>=3.1.0
# reportlab>=4.0.8  # Optional

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.12.0
flake8>=7.0.0
